requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dlib==19.24.2
dnspython==2.4.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.0
MarkupSafe==2.1.3
numpy==1.25.2
Pillow==10.0.0
pinecone-client==2.2.2
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zipp==3.16.2